refactor(mobile): 역 검색 결과 본문 분리
